+++ app/Http/Controllers/Api/AutreImagerieController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Models\Patient;
 use App\Models\AutreImagerie;
 use LogActivity;
+use PatientAccess;
 use Auth;
 
 class AutreImagerieController extends Controller
@@ -18,7 +19,14 @@ class AutreImagerieController extends Controller
      */
     public function index($patient)
     {
-        return AutreImagerie::where('patient_id', $patient)->get();
+        //Check for permission access
+        $permission = PatientAccess::viewPermission($patient);
+        if($permission == 'unauthorized'){
+            return response(['error' => 1, 'message' => 'Patient AutresImagerie data isn\'t available'], 404);
+        }
+
+        $autre_imagerie = AutreImagerie::where('patient_id', $patient)->get();
+        return ['data' =>  $autre_imagerie, 'permission' => $permission];
     }
 
     /**
@@ -29,15 +37,16 @@ class AutreImagerieController extends Controller
      */
     public function store(Request $request, $patient)
     {
-        $patient = Patient::find($patient);
-        if(!$patient){
-            return response(['error' => 1, 'message' => 'Patient doesn\'t exist'], 404);
+        //Check for permission access
+        $permission     = PatientAccess::viewPermission($patient);
+        if($permission != 'write'){
+            return response(['error' => 1, 'message' => 'Not authorized to update patient\'s AutresImagerie data'], 404);
         }
 
+        $patient = Patient::find($patient);
         $autre_imagerie        = new AutreImagerie;
         $autre_imagerie->date  = $request->date;
         $autre_imagerie->value = $request->value;
-
         $patient->autres_imageries()->save($autre_imagerie);
 
         $loggedInUser = $request->user();
@@ -53,7 +62,14 @@ class AutreImagerieController extends Controller
      */
     public function show($patient, $autre_imagerie)
     {
-        return AutreImagerie::find($autre_imagerie);
+        //Check for permission access 
+        $permission = PatientAccess::viewPermission($patient);
+        if($permission == 'unauthorized'){
+            return response(['error' => 1, 'message' => 'Patient Echographie data isn\'t available'], 404);
+        }
+
+        $autre_imagerie = AutreImagerie::find($autre_imagerie);
+        return ['data' =>  $autre_imagerie, 'permission' => $permission];
     }
 
     /**
@@ -65,11 +81,15 @@ class AutreImagerieController extends Controller
      */
     public function update(Request $request, $patient, $autre_imagerie)
     {
+        //Check for permission access
+        $permission     = PatientAccess::viewPermission($patient);
+        if($permission != 'write'){
+            return response(['error' => 1, 'message' => 'Not authorized to update patient\'s Echographie data'], 404);
+        }
+
         $autre_imagerie = AutreImagerie::find($autre_imagerie);
-        
         $autre_imagerie->date  = $request->date    ?? $autre_imagerie->date;
         $autre_imagerie->value = $request->value   ?? $autre_imagerie->value;
-
         $autre_imagerie->update();
 
         $patient        = Patient::find($patient);
@@ -86,6 +106,12 @@ class AutreImagerieController extends Controller
      */
     public function destroy($patient, $autre_imagerie)
     {
+        //Check for permission access
+        $permission     = PatientAccess::viewPermission($patient);
+        if($permission != 'write'){
+            return response(['error' => 1, 'message' => 'Not authorized to update patient\'s Echographie data'], 404);
+        }
+
         $autre_imagerie = AutreImagerie::find($autre_imagerie);
         $autre_imagerie->delete();
