+++ app/Http/Controllers/Api/EvolutionController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Models\Patient;
 use App\Models\Evolution;
 use LogActivity;
+use PatientAccess;
 use Auth;
 
 class EvolutionController extends Controller
@@ -18,7 +19,14 @@ class EvolutionController extends Controller
      */
     public function index($patient)
     {
-        return Evolution::where('patient_id', $patient)->get();
+        //Check for permission access
+        $permission = PatientAccess::viewPermission($patient);
+        if($permission == 'unauthorized'){
+            return response(['error' => 1, 'message' => 'Patient Evolution data isn\'t available'], 404);
+        }
+
+        $evolution = Evolution::where('patient_id', $patient)->get();
+        return ['data' =>  $evolution, 'permission' => $permission];
     }
 
     /**
@@ -29,13 +37,14 @@ class EvolutionController extends Controller
      */
     public function store(Request $request, $patient)
     {
-        $patient = Patient::find($patient);
-        if(!$patient){
-            return response(['error' => 1, 'message' => 'Patient doesn\'t exist'], 404);
+        //Check for permission access
+        $permission     = PatientAccess::viewPermission($patient);
+        if($permission != 'write'){
+            return response(['error' => 1, 'message' => 'Not authorized to update patient\'s Evolution data'], 404);
         }
 
+        $patient = Patient::find($patient);
         $evolution                  = new Evolution;
-
         $evolution->date            = $request->date;
         $evolution->indicateur      = $request->indicateur;
         $evolution->cliniques       = $request->cliniques;
@@ -43,7 +52,6 @@ class EvolutionController extends Controller
         $evolution->traitement      = $request->traitement;
         $evolution->avis            = $request->avis;
         $evolution->conclusion      = $request->conclusion;
-
         $patient->evolutions()->save($evolution);
 
         $loggedInUser = $request->user();
@@ -59,11 +67,18 @@ class EvolutionController extends Controller
      */
     public function find_by_date($patient, $date)
     {
+        //Check for permission access 
+        $permission = PatientAccess::viewPermission($patient);
+        if($permission == 'unauthorized'){
+            return response(['error' => 1, 'message' => 'Patient Evolution data isn\'t available'], 404);
+        }
+
         $array_date     = explode('-', $date);
         $ISO_str_date   = $array_date[2].'-'.$array_date[1].'-'.$array_date[0];  
-        return Evolution::where('patient_id', $patient)
-                        ->where('date', $ISO_str_date)
-                        ->get();
+        $evolution      = Evolution::where('patient_id', $patient)
+                                    ->where('date', $ISO_str_date)
+                                    ->get();
+        return ['data' =>  $evolution, 'permission' => $permission];
     }
 
     /**
@@ -74,7 +89,14 @@ class EvolutionController extends Controller
      */
     public function show($patient, $evolution)
     {
-        return Evolution::find($evolution);
+        //Check for permission access 
+        $permission = PatientAccess::viewPermission($patient);
+        if($permission == 'unauthorized'){
+            return response(['error' => 1, 'message' => 'Patient Evolution data isn\'t available'], 404);
+        }
+
+        $evolution = Evolution::find($evolution);
+        return ['data' =>  $evolution, 'permission' => $permission];
     }
 
     /**
@@ -86,8 +108,13 @@ class EvolutionController extends Controller
      */
     public function update(Request $request, $patient, $evolution)
     {
-        $evolution = Evolution::find($evolution);
+        //Check for permission access
+        $permission     = PatientAccess::viewPermission($patient);
+        if($permission != 'write'){
+            return response(['error' => 1, 'message' => 'Not authorized to update patient\'s Evolution data'], 404);
+        }
 
+        $evolution = Evolution::find($evolution);
         $evolution->date            = $request->date            ?? $evolution->date;
         $evolution->indicateur      = $request->indicateur      ?? $evolution->indicateur;
         $evolution->cliniques       = $request->cliniques       ?? $evolution->cliniques;
@@ -95,7 +122,6 @@ class EvolutionController extends Controller
         $evolution->traitement      = $request->traitement      ?? $evolution->traitement;
         $evolution->avis            = $request->avis            ?? $evolution->avis;
         $evolution->conclusion      = $request->conclusion      ?? $evolution->conclusion;
-     
         $evolution->update();
 
         $patient        = Patient::find($patient);
@@ -112,6 +138,12 @@ class EvolutionController extends Controller
      */
     public function destroy($patient, $evolution)
     {
+        //Check for permission access
+        $permission     = PatientAccess::viewPermission($patient);
+        if($permission != 'write'){
+            return response(['error' => 1, 'message' => 'Not authorized to update patient\'s Evolution data'], 404);
+        }
+
         $evolution = Evolution::find($evolution);
         $evolution->delete();
